Uncomment and relocate billing / receipts test cases.
