+++ integration/node.test.js
@@ -17,13 +17,19 @@ const {
   TX_BYTES_LIMIT,
   BATCH_TX_LIST_SIZE_LIMIT,
   TX_POOL_SIZE_LIMIT_PER_ACCOUNT,
+  MICRO_AIN,
 } = require('../common/constants');
 const CommonUtil = require('../common/common-util');
+const PathUtil = require('../common/path-util');
+const DB = require('../db');
 const {
-  waitUntilTxFinalized,
   parseOrLog,
   setUpApp,
+  waitForNewBlocks,
   waitUntilNetworkIsReady,
+  waitUntilTxFinalized,
+  getLastBlockNumber,
+  getBlockByNumber,
 } = require('../unittest/test-util');
 
 const ENV_VARIABLES = [
@@ -3125,4 +3131,470 @@ describe('Blockchain Node', () => {
       })
     })
   })
+
+  describe('Billing', async () => {
+    let serviceAdmin; // = server1
+    let billingUserA; // = server2
+    let billingUserB; // = server3
+    let userBalancePathA;
+    let userBalancePathB;
+    const billingAccountBalancePathA = '/get_value?ref=/service_accounts/billing/test_billing/A/balance';
+    before(async () => {
+      serviceAdmin =
+          parseOrLog(syncRequest('GET', server1 + '/get_address').body.toString('utf-8')).result;
+      billingUserA =
+          parseOrLog(syncRequest('GET', server2 + '/get_address').body.toString('utf-8')).result;
+      billingUserB =
+          parseOrLog(syncRequest('GET', server3 + '/get_address').body.toString('utf-8')).result;
+      userBalancePathA = `/get_value?ref=/accounts/${billingUserA}/balance`;
+      userBalancePathB = `/get_value?ref=/accounts/${billingUserB}/balance`;
+      const adminConfig = {
+        [serviceAdmin]: true,
+        [billingUserA]: true,
+        [billingUserB]: true
+      };
+      const billingConfig = {
+        A: { 
+          users: {
+            [serviceAdmin]: true,
+            [billingUserA]: true
+          }
+        },
+        B: {
+          users: {
+            [serviceAdmin]: true,
+            [billingUserB]: true
+          }
+        }
+      };
+      await setUpApp('test_billing', serverList, { admin: adminConfig, billing: billingConfig });
+
+      const server4Addr =
+          parseOrLog(syncRequest('GET', server4 + '/get_address').body.toString('utf-8')).result;
+      const transferRes = parseOrLog(syncRequest('POST', server4 + '/set', {json: {
+        op_list: [
+          {
+            ref: `/transfer/${server4Addr}/billing|test_billing|A/${Date.now()}/value`,
+            value: 100,
+            type: 'SET_VALUE'
+          },
+          {
+            ref: `/transfer/${server4Addr}/billing|test_billing|B/${Date.now()}/value`,
+            value: 100,
+            type: 'SET_VALUE'
+          }
+        ],
+        nonce: -1,
+        timestamp: Date.now(),
+      }}).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, transferRes.tx_hash))) {
+        console.error(`Failed to check finalization of transfer tx.`);
+      }
+    });
+
+    it('app txs are not charged by transfer', async () => {
+      // NOTE(platfowner): A pre-tx to guarantee that the service state delta of the next tx
+      // is zero.
+      const txPreRes = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: '/apps/test_billing/test_pre',
+          value: 'testing app tx',
+          gas_price: 1,
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txPreRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      const balanceBefore = parseOrLog(syncRequest('GET', server2 + userBalancePathA).body.toString('utf-8')).result;
+      const txWithoutBillingRes = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: '/apps/test_billing/test',
+          value: 'testing app tx',
+          gas_price: 1,
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txWithoutBillingRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      const balanceAfter = parseOrLog(syncRequest('GET', server2 + userBalancePathA).body.toString('utf-8')).result;
+      assert.deepEqual(balanceAfter, balanceBefore);
+
+      const billingAccountBalanceBefore = parseOrLog(syncRequest(
+          'GET', server2 + billingAccountBalancePathA).body.toString('utf-8')).result;
+      const txWithBillingRes = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: '/apps/test_billing/test',
+          value: 'testing app tx',
+          gas_price: 1,
+          billing: 'test_billing|A',
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txWithBillingRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      const billingAccountBalanceAfter = parseOrLog(syncRequest(
+          'GET', server2 + billingAccountBalancePathA).body.toString('utf-8')).result;
+      assert.deepEqual(billingAccountBalanceAfter, billingAccountBalanceBefore);
+    });
+
+    it('app-dependent service tx: individual account', async () => {
+      const gasPrice = 1;
+      const txRes = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: '/manage_app/test_billing/config/service/staking/lockup_duration',
+          value: 1000,
+          gas_price: gasPrice,
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      // NOTE(liayoo): Checking the gas fee was collected instead of account balances, since the
+      // nodes also participate in consensus & get the collected fees as rewards.
+      const tx = parseOrLog(syncRequest('GET', server2 + `/get_transaction?hash=${txRes.tx_hash}`).body.toString('utf-8')).result;
+      const gasFeeCollected = parseOrLog(syncRequest(
+        'GET',
+        `${server2}/get_value?ref=/gas_fee/collect/${tx.number}/${billingUserA}/${txRes.tx_hash}/amount`
+      ).body.toString('utf-8')).result;
+      assert.deepEqual(
+        gasFeeCollected,
+        gasPrice * MICRO_AIN * txRes.result.gas_amount_charged
+      );
+    });
+
+    it('app-dependent service tx: invalid billing param', async () => {
+      const txResBody = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: '/manage_app/test_billing/config/service/staking/lockup_duration',
+          value: 1000,
+          gas_price: 1,
+          billing: 'A',
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8'));
+      assert.deepEqual(txResBody, {code: 1, result: { tx_hash: null, result: false }});
+    });
+
+    it('app-dependent service tx: not a billing account user', async () => {
+      const txResBody = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: '/manage_app/test_billing/config/service/staking/lockup_duration',
+          value: 1000,
+          gas_price: 1,
+          billing: 'test_billing|B',
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8'));
+      expect(txResBody.code).to.equals(1);
+      expect(txResBody.result.result.code).to.equals(33);
+      expect(txResBody.result.result.error_message).to.equals("[precheckTxBillingParams] User doesn't have permission to the billing account");
+    });
+
+    it('app-dependent service tx: billing account', async () => {
+      const billingAccountBalanceBefore = parseOrLog(syncRequest(
+        'GET', server2 + billingAccountBalancePathA).body.toString('utf-8')).result;
+      const gasPrice = 1;
+      const txRes = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: '/manage_app/test_billing/config/service/staking/lockup_duration',
+          value: 1000,
+          gas_price: 1,
+          billing: 'test_billing|A',
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      const billingAccountBalanceAfter = parseOrLog(syncRequest(
+          'GET', server2 + billingAccountBalancePathA).body.toString('utf-8')).result;
+      assert.deepEqual(
+        billingAccountBalanceAfter,
+        billingAccountBalanceBefore - (gasPrice * MICRO_AIN * txRes.result.gas_amount_charged)
+      );
+    });
+
+    it('app-independent service tx: individual account', async () => {
+      const gasPrice = 1;
+      const txRes = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: `/transfer/${billingUserA}/${billingUserB}/${Date.now()}/value`,
+          value: 1,
+          gas_price: gasPrice,
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      // NOTE(liayoo): Checking the gas fee was collected instead of account balances, since the
+      // nodes also participate in consensus & get the collected fees as rewards.
+      const tx = parseOrLog(syncRequest('GET', server2 + `/get_transaction?hash=${txRes.tx_hash}`).body.toString('utf-8')).result;
+      const gasFeeCollected = parseOrLog(syncRequest(
+        'GET',
+        `${server2}/get_value?ref=/gas_fee/collect/${tx.number}/${billingUserA}/${txRes.tx_hash}/amount`
+      ).body.toString('utf-8')).result;
+      assert.deepEqual(
+        gasFeeCollected,
+        gasPrice * MICRO_AIN * txRes.result.gas_amount_charged
+      );
+    });
+
+    it('app-independent service tx: billing account', async () => {
+      const billingAccountBalanceBefore = parseOrLog(syncRequest(
+          'GET', server2 + billingAccountBalancePathA).body.toString('utf-8')).result;
+      const gasPrice = 1;
+      const txRes = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+          ref: `/transfer/${billingUserA}/${billingUserB}/${Date.now()}/value`,
+          value: 1,
+          gas_price: gasPrice,
+          billing: 'test_billing|A',
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      const billingAccountBalanceAfter = parseOrLog(syncRequest(
+          'GET', server2 + billingAccountBalancePathA).body.toString('utf-8')).result;
+      assert.deepEqual(
+        billingAccountBalanceAfter,
+        billingAccountBalanceBefore - (gasPrice * MICRO_AIN * txRes.result.gas_amount_charged)
+      );
+    });
+
+    it('multi-set service tx: individual account', async () => {
+      const gasPrice = 1;
+      const txRes = parseOrLog(syncRequest('POST', server2 + '/set', {json: {
+          op_list: [
+            {
+              ref: `/transfer/${billingUserA}/${billingUserB}/${Date.now()}/value`,
+              value: 1,
+              type: 'SET_VALUE'
+            },
+            {
+              ref: `/manage_app/test_billing/config/service/staking/lockup_duration`,
+              value: 100,
+              type: 'SET_VALUE'
+            }
+          ],
+          gas_price: gasPrice,
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      // NOTE(liayoo): Checking the gas fee was collected instead of account balances, since the
+      // nodes also participate in consensus & get the collected fees as rewards.
+      const tx = parseOrLog(syncRequest('GET', server2 + `/get_transaction?hash=${txRes.tx_hash}`).body.toString('utf-8')).result;
+      const gasFeeCollected = parseOrLog(syncRequest(
+        'GET',
+        `${server2}/get_value?ref=/gas_fee/collect/${tx.number}/${billingUserA}/${txRes.tx_hash}/amount`
+      ).body.toString('utf-8')).result;
+      assert.deepEqual(
+        gasFeeCollected,
+        gasPrice * MICRO_AIN * txRes.result.gas_amount_charged
+      );
+    });
+
+    it('multi-set service tx: billing account', async () => {
+      const billingAccountBalanceBefore = parseOrLog(syncRequest(
+          'GET', server2 + billingAccountBalancePathA).body.toString('utf-8')).result;
+      const gasPrice = 1;
+      const txRes = parseOrLog(syncRequest('POST', server2 + '/set', {json: {
+          op_list: [
+            {
+              ref: `/transfer/${billingUserA}/${billingUserB}/${Date.now()}/value`,
+              value: 1,
+              type: 'SET_VALUE'
+            },
+            {
+              ref: `/manage_app/test_billing/config/service/staking/lockup_duration`,
+              value: 100,
+              type: 'SET_VALUE'
+            }
+          ],
+          billing: 'test_billing|A',
+          gas_price: gasPrice,
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, txRes.tx_hash))) {
+        console.error(`Failed to check finalization of app tx.`);
+      }
+      const billingAccountBalanceAfter = parseOrLog(syncRequest(
+          'GET', server2 + billingAccountBalancePathA).body.toString('utf-8')).result;
+      assert.deepEqual(
+        billingAccountBalanceAfter,
+        billingAccountBalanceBefore - (gasPrice * MICRO_AIN * txRes.result.gas_amount_charged)
+      );
+    });
+
+    it('multi-set service tx: multiple apps', async () => {
+      // Set up another app
+      const appStakingRes = parseOrLog(syncRequest('POST', server1 + '/set_value', {json: {
+        ref: `/staking/test_billing_2/${serviceAdmin}/0/stake/${Date.now()}/value`,
+        value: 1
+      }}).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, appStakingRes.tx_hash))) {
+        console.error(`Failed to check finalization of app staking tx.`);
+      }
+
+      const createAppRes = parseOrLog(syncRequest('POST', server2 + '/set_value', {json: {
+        ref: '/manage_app/test_billing_2/create/0',
+        value: {
+          admin: {
+            [serviceAdmin]: true,
+            [billingUserA]: true,
+            [billingUserB]: true
+          },
+          billing: {
+            '0': { 
+              users: {
+                [serviceAdmin]: true,
+                [billingUserA]: true,
+                [billingUserB]: true
+              }
+            }
+          }
+        },
+        nonce: -1,
+        timestamp: Date.now(),
+      }}).body.toString('utf-8')).result;
+      if (!(await waitUntilTxFinalized(serverList, createAppRes.tx_hash))) {
+        console.error(`Failed to check finalization of create app tx.`);
+      }
+
+      const txResBody = parseOrLog(syncRequest('POST', server1 + '/set', {json: {
+          op_list: [
+            {
+              ref: `/manage_app/test_billing/config/service/staking/lockup_duration`,
+              value: 100,
+              type: 'SET_VALUE'
+            },
+            {
+              ref: `/manage_app/test_billing_2/config/service/staking/lockup_duration`,
+              value: 100,
+              type: 'SET_VALUE'
+            }
+          ],
+          billing: 'test_billing|A',
+          gas_price: 1,
+          nonce: -1,
+          timestamp: Date.now(),
+        }
+      }).body.toString('utf-8'));
+      assert.deepEqual(txResBody.result.result, {
+        "bandwidth_gas_amount": 0,
+        "error_message": "[precheckTxBillingParams] Multiple app-dependent service operations for a billing account",
+        "code": 16
+      });
+    });
+  });
+
+  describe('Tx Receipts', () => {
+    it(`records a transaction receipt`, async () => {
+      const txSignerAddress = parseOrLog(syncRequest(
+          'GET', server1 + '/get_address').body.toString('utf-8')).result;
+      const request = {
+        ref: '/apps/test/test_value/some/path',
+        value: "some value"
+      };
+      const body = parseOrLog(syncRequest(
+          'POST', server1 + '/set_value', {json: request}).body.toString('utf-8'));
+      assert.deepEqual(_.get(body, 'result.result.code'), 0);
+      expect(_.get(body, 'result.tx_hash')).to.not.equal(null);
+      expect(body.code).to.equal(0);
+      if (!(await waitUntilTxFinalized(serverList, _.get(body, 'result.tx_hash')))) {
+        console.error(`Failed to check finalization of tx.`);
+      }
+
+      const receipt = parseOrLog(syncRequest(
+          'GET', server1 + `/get_value?ref=${PathUtil.getReceiptPath(body.result.tx_hash)}`)
+          .body.toString('utf-8')).result;
+      assert.deepEqual(receipt.address, txSignerAddress);
+      assert.deepEqual(receipt.exec_result, {
+        "code": 0,
+        "gas_amount_charged": 8,
+        "gas_cost_total": 0
+      });
+    });
+
+    it(`removes an old transaction receipt`, async () => {
+      const MAX_BLOCK_NUMBERS_FOR_RECEIPTS = 100;
+      let lastBlockNumber = getLastBlockNumber(server1);
+      if (lastBlockNumber <= MAX_BLOCK_NUMBERS_FOR_RECEIPTS) {
+        await waitForNewBlocks(server1, MAX_BLOCK_NUMBERS_FOR_RECEIPTS - lastBlockNumber + 1);
+        lastBlockNumber = getLastBlockNumber(server1);
+      }
+      let oldBlockNumber = lastBlockNumber - MAX_BLOCK_NUMBERS_FOR_RECEIPTS;
+      let oldBlock = getBlockByNumber(server1, oldBlockNumber);
+      while (!oldBlock.transactions.length) {
+        oldBlock = getBlockByNumber(server1, --oldBlockNumber);
+        await CommonUtil.sleep(2000);
+      }
+      for (const tx of oldBlock.transactions) {
+        const receipt = parseOrLog(syncRequest(
+          'GET', server1 + `/get_value?ref=${PathUtil.getReceiptPath(tx.hash)}`)
+          .body.toString('utf-8')).result;
+        assert.deepEqual(receipt, null);
+      }
+    });
+
+    it('failed transaction', async () => {
+      const server1Address = parseOrLog(syncRequest(
+        'GET', server1 + '/get_address').body.toString('utf-8')).result;
+      const server2Address = parseOrLog(syncRequest(
+        'GET', server2 + '/get_address').body.toString('utf-8')).result;
+      const failingTx = {
+        ref: `/transfer/${server1Address}/${server2Address}/${Date.now()}/value`,
+        value: 10000000000,
+        gas_price: 1
+      }
+      const body = parseOrLog(syncRequest(
+        'POST', server1 + '/set_value', {json: failingTx}).body.toString('utf-8'));
+      assert.deepEqual(body.result.result.code, 103);
+      assert.deepEqual(body.result.result.bandwidth_gas_amount, 1);
+      assert.deepEqual(body.result.result.gas_amount_total, {
+        "bandwidth": {
+          "service": 1
+        },
+        "state": {
+          "service": 0
+        }
+      });
+      assert.deepEqual(body.result.result.gas_cost_total, 0.000001);
+      
+      if (!(await waitUntilTxFinalized(serverList, _.get(body, 'result.tx_hash')))) {
+        console.error(`Failed to check finalization of tx.`);
+      }
+
+      // Failed tx's receipt is in state
+      const txHash = body.result.tx_hash;
+      const receipt = parseOrLog(syncRequest(
+        'GET', server2 + `/get_value?ref=${PathUtil.getReceiptPath(txHash)}`).body.toString('utf-8')).result;
+      expect(receipt).to.not.equal(null);
+      assert.deepEqual(receipt.exec_result, DB.trimExecutionResult(body.result.result));
+
+      // Failed tx's gas fees have been collected
+      const blockNumber = receipt.block_number;
+      const gasFeeCollected = parseOrLog(syncRequest(
+        'GET', server2 + `/get_value?ref=/gas_fee/collect/${blockNumber}/${server1Address}/${txHash}/amount`
+      ).body.toString('utf-8')).result;
+      assert.deepEqual(gasFeeCollected, body.result.result.gas_cost_total);
+
+      // Failed tx is in a block
+      const block = getBlockByNumber(server2, blockNumber);
+      expect(block).to.not.equal(undefined);
+      expect(block.transactions.find((tx) => tx.hash === txHash)).to.not.equal(undefined);
+    });
+  });
 });
